Add 1.20.4 adapter
