--- internal/v1_21_R1/pom.xml
@@ -1,77 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-  ~ Copyright (C) 2011-2023 lishid. All rights reserved.
-  ~
-  ~ This program is free software: you can redistribute it and/or modify
-  ~ it under the terms of the GNU General Public License as published by
-  ~ the Free Software Foundation, version 3.
-  ~
-  ~ This program is distributed in the hope that it will be useful,
-  ~ but WITHOUT ANY WARRANTY; without even the implied warranty of
-  ~ MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-  ~ GNU General Public License for more details.
-  ~
-  ~ You should have received a copy of the GNU General Public License
-  ~ along with this program. If not, see <http://www.gnu.org/licenses/>.
-  -->
-
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-    xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-
-  <parent>
-    <artifactId>openinvparent</artifactId>
-    <groupId>com.lishid</groupId>
-    <relativePath>../../pom.xml</relativePath>
-    <version>5.1.4-SNAPSHOT</version>
-  </parent>
-
-  <artifactId>openinvadapter1_21_R1</artifactId>
-  <name>OpenInvAdapter1_21_R1</name>
-
-  <properties>
-    <maven.compiler.system>21</maven.compiler.system>
-    <maven.compiler.release>21</maven.compiler.release>
-    <spigot.version>1.21.1-R0.1-SNAPSHOT</spigot.version>
-  </properties>
-
-  <dependencies>
-    <dependency>
-      <groupId>org.spigotmc</groupId>
-      <artifactId>spigot-api</artifactId>
-      <version>${spigot.version}</version>
-    </dependency>
-    <dependency>
-      <artifactId>spigot</artifactId>
-      <groupId>org.spigotmc</groupId>
-      <scope>provided</scope>
-      <version>${spigot.version}</version>
-      <classifier>remapped-mojang</classifier>
-    </dependency>
-    <dependency>
-      <artifactId>openinvapi</artifactId>
-      <groupId>com.lishid</groupId>
-    </dependency>
-    <dependency>
-      <artifactId>openinvcommon</artifactId>
-      <groupId>com.lishid</groupId>
-    </dependency>
-    <dependency>
-      <artifactId>annotations</artifactId>
-      <groupId>org.jetbrains</groupId>
-    </dependency>
-  </dependencies>
-
-  <build>
-    <plugins>
-      <plugin>
-        <artifactId>maven-compiler-plugin</artifactId>
-      </plugin>
-      <plugin>
-        <groupId>net.md-5</groupId>
-        <artifactId>specialsource-maven-plugin</artifactId>
-      </plugin>
-    </plugins>
-  </build>
-
-</project>
