--- internal/v1_21_R1/src/main/java/com/lishid/openinv/internal/v1_21_R1/container/bukkit/OpenPlayerInventorySelf.java
@@ -1,26 +1,0 @@
-package com.lishid.openinv.internal.v1_21_R1.container.bukkit;
-
-import com.lishid.openinv.internal.v1_21_R1.container.OpenInventory;
-import org.bukkit.inventory.ItemStack;
-import org.jetbrains.annotations.NotNull;
-
-public class OpenPlayerInventorySelf extends OpenPlayerInventory {
-
-  private final int offset;
-
-  public OpenPlayerInventorySelf(@NotNull OpenInventory inventory, int offset) {
-    super(inventory);
-    this.offset = offset;
-  }
-
-  @Override
-  public ItemStack getItem(int index) {
-    return super.getItem(offset + index);
-  }
-
-  @Override
-  public void setItem(int index, ItemStack item) {
-    super.setItem(offset + index, item);
-  }
-
-}
