+++ plugin/src/main/java/com/lishid/openinv/util/InternalAccessor.java
@@ -40,10 +40,10 @@ public class InternalAccessor {
     public InternalAccessor(@NotNull Logger logger, @NotNull LanguageManager lang) {
 
         try {
-            if (BukkitVersions.MINECRAFT.equals(Version.of(1, 21, 3))) {
+            if (BukkitVersions.MINECRAFT.equals(Version.of(1, 21, 4))) {
+                internal = new com.lishid.openinv.internal.v1_21_R3.InternalAccessor(logger, lang);
+            } else if (BukkitVersions.MINECRAFT.equals(Version.of(1, 21, 3))) {
                 internal = new com.lishid.openinv.internal.v1_21_R2.InternalAccessor(logger, lang);
-            } else if (BukkitVersions.MINECRAFT.equals(Version.of(1, 21, 1))) {
-                internal = new com.lishid.openinv.internal.v1_21_R1.InternalAccessor(logger, lang);
             }
             if (internal != null) {
                 InventoryAccess.setProvider(internal::get);
@@ -123,6 +123,9 @@ public class InternalAccessor {
         }
         if (BukkitVersions.MINECRAFT.lessThanOrEqual(Version.of(1, 21))) { // 1.20.4, 1.20.6, 1.21
             return "https://github.com/Jikoo/OpenInv/releases/tag/5.1.2";
+        }
+        if (BukkitVersions.MINECRAFT.lessThanOrEqual(Version.of(1, 21, 2))) {
+            return "https://github.com/Jikoo/OpenInv/releases/tag/5.1.3";
         }
         return "https://github.com/Jikoo/OpenInv/releases";
     }
